Don't try this at home, kids: dirty hack to add support for changing server accounts
Server accounts UUID is just that, a UUID. And yet we're using it for
authentication anyways ¯\_(ツ)_/¯

I realize that we should generate a secret string of bytes or something,
but there's just no time.

Furthermore, simply setting Caladrius.user.sAcc doesn't seem to be
working and I'm too lazy to figure it out. So now I'm doing the correct
thing and actually passing the data to the classes that need it. Buuut
I don't have the time to bother updating all the classes, so now we're
stuck in this nasty inbetween state. Yuck.